refactor: Separate parser functions by modules
